feat: sort menu items alphabetically (#1823)
