+++ src/renderer/store/modules/plugin.js
@@ -1,6 +1,7 @@
 import Vue from 'vue'
 import pluginManager from '@/services/plugin-manager'
 import releaseService from '@/services/release'
+import { sortByProps } from '@/utils'
 import { cloneDeep, uniqBy } from 'lodash'
 import semver from 'semver'
 
@@ -212,7 +213,7 @@ export default {
         menuItems.push(...plugin.menuItems)
       }
 
-      return menuItems
+      return [...menuItems].sort(sortByProps('title'))
     },
 
     themes: (_, getters) => {
