+++ src/renderer/utils/index.js
@@ -1,30 +1,32 @@
+import sortByProps from './sort-by-props'
+
 /**
  * Converts the first character of `string` to upper case
  * @param {String} str
  * @return {String} Return the converted string
  */
-export const upperFirst = str => `${str.charAt(0).toUpperCase()}${str.slice(1)}`
+const upperFirst = str => `${str.charAt(0).toUpperCase()}${str.slice(1)}`
 
 /**
  * Converts the first character of `string` to upper case and the remaining to lower case
  * @param {String} str
  * @return {String} Return the capitalized string
  */
-export const capitalize = str => `${str.charAt(0).toUpperCase()}${str.slice(1).toLowerCase()}`
+const capitalize = str => `${str.charAt(0).toUpperCase()}${str.slice(1).toLowerCase()}`
 
 /**
  * Checks if `value` is `null` or `undefined`
  * @param {*} val
  * @return {Boolean} Return `true` if `value` is nullish, else `false`
  */
-export const isNil = val => val === null || val === undefined
+const isNil = val => val === null || val === undefined
 
 /**
  * Computes the minimum value of `array`
  * @param {Array} nums
  * @return {*} Return the minimum value
  */
-export const min = nums => {
+const min = nums => {
   if (nums.length) {
     return Math.min(...nums)
   }
@@ -35,8 +37,17 @@ export const min = nums => {
  * @param {Array} nums
  * @return {*} Return the maximum value
  */
-export const max = nums => {
+const max = nums => {
   if (nums.length) {
     return Math.max(...nums)
   }
+}
+
+export {
+  upperFirst,
+  capitalize,
+  isNil,
+  min,
+  max,
+  sortByProps
 }
